+++ components/organisms/CronDropDown.tsx
@@ -1,6 +1,5 @@
 import useCreateBucketStore from '@/contexts/useCreateBucketStore';
 import { IoIosArrowDown } from 'react-icons/io';
-import { useState } from 'react';
 import { cn } from '@/lib/utils';
 import ColorChip from '../atoms/ColorChips';
 import {
